set login as startpage, remove console.logs
