+++ pages/index.js
@@ -1,42 +1,21 @@
-import CatList from "@/components/CatList/CatList";
-import Navigation from "@/components/Navigation/Navigation";
-import Link from "next/link";
+import Login from "@/components/Login/Login";
 import styled from "styled-components";
 
-export default function HomePage({ catList }) {
+export default function LoginPage() {
   return (
     <>
-      <StyledContainer>
-        <StyledHeading>Your Cats</StyledHeading>
-        <StyledLink href={"/addcat"}>Add New Cat</StyledLink>
-        <CatList catList={catList} />
-      </StyledContainer>
-      <Navigation />
+      <StyledLoginBackground>
+        <Login />
+      </StyledLoginBackground>
     </>
   );
 }
 
-const StyledHeading = styled.h1`
-  margin-left: 5%;
-  text-align: center;
-  color: white;
-  letter-spacing: 2.5px;
-  text-shadow: 2px 2px 10px #f0caa3;
-`;
-
-const StyledLink = styled(Link)`
-  max-width: 37%;
-  background: transparent;
-  backdrop-filter: blur(2px);
-  color: #f0caa3;
-  font-size: 1em;
-  margin: 1em;
-  padding: 0.25em 1em;
-  border: 2px solid #f0caa3;
-  border-radius: 3px;
-  text-decoration: none;
-`;
-
-const StyledContainer = styled.div`
-  padding-bottom: 60px;
+const StyledLoginBackground = styled.div`
+  background-image: url("/login_background2.png");
+  background-size: cover;
+  background-repeat: no-repeat;
+  background-position: center;
+  min-height: 100vh;
+  width: 100%;
 `;
